cmd: rename MonitorLoop and move common to internal
